% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivaleff.R
\name{chisq_fisher}
\alias{chisq_fisher}
\title{chisq_fisher}
\usage{
chisq_fisher(
  data,
  ncol,
  row_names = NULL,
  col_names = NULL,
  digits = 2,
  p_digits = 3
)
}
\arguments{
\item{data}{A numeric vector representing contingency table counts, filled row-wise.}

\item{ncol}{Integer. Number of columns in the contingency table.}

\item{row_names}{Optional character vector for row labels.}

\item{col_names}{Optional character vector for column labels.}

\item{language}{Output language. Either "en" (default) or "cn".}
}
\value{
A list including the contingency table, expected values, p-value, method used and other metadata.
}
\description{
Performs Chi-squared or Fisher's exact test based on user-specified contingency table data.
Automatically detects whether to use Yates correction or Fisher's test.
}
\details{
Used to quickly calculate chisq or fisher values
}
\examples{
chisq_fisher(
  data = c(15, 25, 10, 13),
  ncol = 2
)
}
\author{
camellia
}
