% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivaleff.R
\name{ORR_DCR}
\alias{ORR_DCR}
\title{ORR_DCR}
\usage{
ORR_DCR(x1, x2, n, digits = 2)
}
\arguments{
\item{x1}{the number of ORR events.}

\item{x2}{the number of DCR events.}

\item{n}{the Total.}
}
\value{
A part of input data.
}
\description{
Used to quickly calculate ORR, DCR values
}
\examples{
library(DescTools)
ORR_DCR(34,40,75)
}
\author{
camellia
}
