% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivaleff.R
\name{summary_variable}
\alias{summary_variable}
\title{summary_variable}
\usage{
summary_variable(data, cols, digits = 2, p_digits = 3)
}
\arguments{
\item{data}{data object.}

\item{variable}{single variable.}
}
\value{
A part of input data.
}
\description{
According to the provided single variable, automatically identify whether it is a categorical variable
or a continuous variable. If it is a continuous variable,
provide the basic characteristics of this variable and conduct a normality test.
If it is a categorical variable, provide the frequency and percentage.
}
\examples{
set.seed(123)
dt <- data.frame(
age = sample(20:80,100,replace = TRUE),
group = sample(c("A", "B"),100,replace = TRUE))
summary_variable(data = dt, col = "age", digits = 2,p_digits=3)
summary_variable(data = dt, col = "group", digits = 2,p_digits=3)
}
\author{
camellia
}
