% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivaleff.R
\name{survival_analysis_by_group}
\alias{survival_analysis_by_group}
\title{survival_analysis_by_group}
\usage{
survival_analysis_by_group(
  data,
  time_var,
  status_var,
  group_var,
  digits = 2,
  p_digits = 3
)
}
\arguments{
\item{data}{data object.}

\item{time_var}{time variable.}

\item{status_var}{Numeric type, only 0/1,0 is censored and 1 is the occurrence of an event.}

\item{group_var}{subgroup variable..}
}
\value{
A part of input data.
}
\description{
When conducting subgroup analysis, the model characteristics can be quickly obtained,
including the median survival data of each group,
subgroup results, HR value, and Log-rank test results.
}
\examples{
library(survival)
set.seed(123)
dt <- data.frame(
pfs1 = c(5, 10, 15, 20, 30, 35, 40, 50, 60, 70),
pfs1_status = c(1, 1, 0, 1, 0, 1, 1, 0, 1, 1),
肝转移 = factor(c("Yes", "No", "Yes", "No", "Yes", "No", "Yes", "No", "Yes", "No")))
result <- survival_analysis_by_group(dt, "pfs1", "pfs1_status", "肝转移")
result
}
\author{
camellia
}
