% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivaleff.R
\name{survival_overview}
\alias{survival_overview}
\title{survival_overview}
\usage{
survival_overview(
  data,
  time_var,
  status_var,
  time_points = c(6, 12),
  digits = 2
)
}
\arguments{
\item{data}{data object.}

\item{time_var}{time variable.}

\item{status_var}{Numeric type, only 0/1,0 is censored and 1 is the occurrence of an event.}

\item{time_points}{Specific 3 time points.}

\item{digits}{Retain the number of decimal places, default is 2}
}
\value{
A part of input data.
}
\description{
When drawing the overall survival curve, the median time and 95\% CI confidence
interval can be quickly viewed and the survival rates
at the specified times can be output.
}
\examples{
library(survival)
set.seed(123)
n <- 100
dt <- data.frame(
pfs1 = rexp(n, rate = 0.1),
pfs1_status = sample(c(0, 1), n, replace = TRUE))
survival_overview(data = dt, time_var = "pfs1", status_var = "pfs1_status", time_points = c(6, 12, 24))
}
\author{
camellia
}
